set defconfigs for every hosts

+++ modules/kernel/lib/kernel-wrapper.nix
@@ -12,7 +12,12 @@
   };
 
   kernel.lib.kernel-wrapper =
-    pkgs: class: dtbMake:
+    {
+      pkgs,
+      defconfig ? null,
+      class ? "x86",
+      dtbMake ? "",
+    }:
     pkgs.stdenvNoCC.mkDerivation {
       name = "kernel-wrapper";
       src = inputs.linux;
@@ -22,7 +27,7 @@
       postPatch = ''
         sed -i 's/static int unprivileged_userns_clone = 1;/extern int unprivileged_userns_clone;/' kernel/fork.c
         ${
-          if (class != "x86") then
+          if (class != "x86" && dtbMake != "") then
             ''
               find "./arch/arm64/boot/dts" -mindepth 1 -maxdepth 1 -type d ! -name "${class}" -exec rm -rf {} +
               cat <<EOF > "./arch/arm64/boot/dts/Makefile"
@@ -37,7 +42,12 @@
               install -Dm644 ${inputs.linux-config}/linux-cachyos/config arch/x86/configs/cachyos_defconfig
             ''
         }
-
+        ${
+          if (class != "x86" && defconfig != null) then
+            "install -Dm644 ${defconfig} arch/arm64/configs/armcust_defconfig"
+          else
+            "cp arch/arm64/configs/defconfig arch/arm64/configs/armcust_defconfig"
+        }
         ${lib.optionalString (class == "allwinner") ''
           mkdir -p "./drivers/net/wireless/uwe5622"
           cp -R "${inputs.uwe5622}''${uwe5622ver#*:}"/{tty-sdio,unisocwcn,unisocwifi,Kconfig,Makefile} "./drivers/net/wireless/uwe5622"
